move model queries out of emit
- colocate state, actor, entry, and route lookups with Model
- reuse the centralized entry lookup for template-use collection
