+++ src/emit.rs
@@ -201,103 +201,6 @@ impl<'a> Model<'a> {
         Ok(model)
     }
 
-    fn state(&self, name: &str) -> Result<&StateDecl> {
-        self.states
-            .get(name)
-            .copied()
-            .or_else(|| self.linked_states.get(name))
-            .ok_or_else(|| ArgentError::new(format!("unknown state `{name}`")))
-    }
-
-    fn storage_state_name(&self, name: &str) -> Result<String> {
-        let state = self.state(name)?;
-        Ok(state.expansion.as_ref().map_or_else(|| name.to_string(), |expansion| expansion.base.clone()))
-    }
-
-    fn storage_state(&self, name: &str) -> Result<&StateDecl> {
-        self.state(&self.storage_state_name(name)?)
-    }
-
-    fn has_state(&self, name: &str) -> bool {
-        self.states.contains_key(name) || self.linked_states.contains_key(name)
-    }
-
-    fn all_states(&self) -> impl Iterator<Item = &StateDecl> {
-        self.states.values().copied().chain(self.linked_states.values())
-    }
-
-    fn actor(&self, name: &str) -> Result<&ActorDecl> {
-        self.actors_by_name
-            .get(name)
-            .copied()
-            .or_else(|| self.linked_actor_decls.get(name))
-            .ok_or_else(|| ArgentError::new(format!("unknown actor `{name}`")))
-    }
-
-    fn actor_model(&self, actor: &str) -> Result<&ActorModel<'a>> {
-        self.actor_models.get(actor).ok_or_else(|| ArgentError::new(format!("unknown app actor `{actor}`")))
-    }
-
-    fn entry_model(&self, actor: &ActorDecl, entry: &EntryDecl) -> Result<&EntryModel<'a>> {
-        self.actor_model(&actor.name)?
-            .entry(&entry.name)
-            .ok_or_else(|| ArgentError::new(format!("unknown entry model `{}::{}`", actor.name, entry.name)))
-    }
-
-    fn linked_actor(&self, name: &str) -> Option<&LinkedActor> {
-        self.linked_actors.get(name)
-    }
-
-    fn actor_state(&self, name: &str) -> Result<&StateDecl> {
-        let actor = self.actor(name)?;
-        self.storage_state(&actor.state)
-    }
-
-    fn route_family_for_actor(&self, actor: &str) -> Option<&RouteFamily> {
-        self.route_families.iter().find(|family| family.actors.iter().any(|member| member == actor))
-    }
-
-    fn route_family(&self, family_id: &str) -> Option<&RouteFamily> {
-        self.route_families.iter().find(|family| family.id == family_id)
-    }
-
-    fn route_transition(&self, source: &str, target: &str) -> Option<&CompilerRouteTransition> {
-        self.route_transitions.get(&(source.to_string(), target.to_string()))
-    }
-
-    fn route_families_for_state(&self, state: &str) -> Vec<&RouteFamily> {
-        self.route_families.iter().filter(|family| family.state == state).collect()
-    }
-
-    fn leader_for(&self, actor: &str) -> &[EntryRefArtifact] {
-        self.leader_for.get(actor).map(Vec::as_slice).unwrap_or(&[])
-    }
-
-    fn is_leader_actor(&self, actor: &str) -> bool {
-        !self.leader_for(actor).is_empty()
-    }
-
-    fn template_selectors_for_entry(&self, actor: &ActorDecl, entry: &EntryDecl) -> Result<BTreeMap<String, TemplateSelector>> {
-        Ok(self.entry_model(actor, entry)?.template_selectors().clone())
-    }
-
-    fn is_actor_enum_type(&self, ty: &TypeRef) -> bool {
-        ty.array.is_none() && self.actor_enums.contains_key(&ty.name)
-    }
-
-    fn expand_actor_refs(&self, refs: &[String]) -> Vec<String> {
-        refs.iter()
-            .flat_map(|actor| {
-                self.actor_enums.get(actor).map_or_else(|| vec![actor.clone()], |actor_enum| actor_enum.variants.clone())
-            })
-            .collect()
-    }
-
-    fn route_targets(&self, actor: &ActorDecl, entry: &EntryDecl, route: &RouteCall) -> Result<Vec<String>> {
-        let selectors = self.template_selectors_for_entry(actor, entry)?;
-        Ok(selectors.get(&route.actor).map_or_else(|| vec![route.actor.clone()], TemplateSelector::route_actors))
-    }
-
     fn validate(&self) -> Result<()> {
         self.validate_reserved_identifiers()?;
         self.validate_state_expansions()?;
